Extract FileAsMediaAttachment for reuse in avatar/header upload
